move rest deadlines classes to new package
