+++ app/src/main/java/org/stepik/android/cache/personal_deadlines/DeadlinesCacheDataSourceImpl.kt
@@ -3,13 +3,13 @@ package org.stepik.android.cache.personal_deadlines
 import io.reactivex.Completable
 import io.reactivex.Maybe
 import io.reactivex.Single
-import org.stepic.droid.features.deadlines.model.Deadline
+import org.stepik.android.domain.personal_deadlines.model.Deadline
 import org.stepik.android.cache.personal_deadlines.model.DeadlineEntity
-import org.stepic.droid.features.deadlines.model.DeadlinesWrapper
+import org.stepik.android.domain.personal_deadlines.model.DeadlinesWrapper
 import org.stepik.android.data.personal_deadlines.source.DeadlinesCacheDataSource
 import org.stepik.android.cache.personal_deadlines.structure.DbStructureDeadlines
 import org.stepik.android.cache.personal_deadlines.dao.PersonalDeadlinesDao
-import org.stepic.droid.features.deadlines.util.getKindOfRecord
+import org.stepik.android.data.personal_deadlines.getKindOfRecord
 import org.stepic.droid.util.AppConstants
 import org.stepic.droid.web.storage.model.StorageRecord
 import org.stepik.android.cache.personal_deadlines.mapper.DeadlineEntityMapper
