+++ app/src/main/java/org/stepik/android/data/personal_deadlines/DeadlinesHelper.kt
@@ -1,3 +1,3 @@
-package org.stepic.droid.features.deadlines.util
+package org.stepik.android.data.personal_deadlines
 
 fun getKindOfRecord(courseId: Long) = "deadline_$courseId"
