+++ app/src/main/java/org/stepik/android/data/personal_deadlines/repository/DeadlinesRepositoryImpl.kt
@@ -4,7 +4,7 @@ import io.reactivex.Completable
 import io.reactivex.Maybe
 import io.reactivex.Single
 import org.stepic.droid.web.storage.model.StorageRecord
-import org.stepic.droid.features.deadlines.model.DeadlinesWrapper
+import org.stepik.android.domain.personal_deadlines.model.DeadlinesWrapper
 import org.stepik.android.data.personal_deadlines.source.DeadlinesCacheDataSource
 import org.stepic.droid.util.doCompletableOnSuccess
 import org.stepic.droid.util.then
