+++ app/src/main/java/org/stepik/android/data/personal_deadlines/source/DeadlinesCacheDataSource.kt
@@ -4,7 +4,7 @@ import io.reactivex.Completable
 import io.reactivex.Maybe
 import io.reactivex.Single
 import org.stepik.android.cache.personal_deadlines.model.DeadlineEntity
-import org.stepic.droid.features.deadlines.model.DeadlinesWrapper
+import org.stepik.android.domain.personal_deadlines.model.DeadlinesWrapper
 import org.stepic.droid.web.storage.model.StorageRecord
 
 interface DeadlinesCacheDataSource {
