+++ app/src/main/java/org/stepik/android/domain/personal_deadlines/model/Deadline.kt
@@ -1,4 +1,4 @@
-package org.stepic.droid.features.deadlines.model
+package org.stepik.android.domain.personal_deadlines.model
 
 import android.os.Parcel
 import android.os.Parcelable
@@ -19,7 +19,11 @@ class Deadline(
     override fun describeContents(): Int = 0
 
     companion object CREATOR : Parcelable.Creator<Deadline> {
-        override fun createFromParcel(parcel: Parcel): Deadline = Deadline(parcel.readLong(), Date(parcel.readLong()))
+        override fun createFromParcel(parcel: Parcel): Deadline =
+            Deadline(
+                parcel.readLong(),
+                Date(parcel.readLong())
+            )
         override fun newArray(size: Int): Array<Deadline?> = arrayOfNulls(size)
     }
 }
