+++ app/src/main/java/org/stepik/android/presentation/course_content/mapper/CourseContentStateMapper.kt
@@ -1,7 +1,7 @@
 package org.stepik.android.presentation.course_content.mapper
 
 import org.stepic.droid.R
-import org.stepic.droid.features.deadlines.model.DeadlinesWrapper
+import org.stepik.android.domain.personal_deadlines.model.DeadlinesWrapper
 import org.stepic.droid.web.storage.model.StorageRecord
 import org.stepik.android.model.Course
 import org.stepik.android.presentation.course_content.CourseContentView
