+++ app/src/main/java/org/stepik/android/remote/personal_deadlines/DeadlinesRemoteDataSourceImpl.kt
@@ -3,8 +3,8 @@ package org.stepik.android.remote.personal_deadlines
 import io.reactivex.Completable
 import io.reactivex.Maybe
 import io.reactivex.Single
-import org.stepic.droid.features.deadlines.model.DeadlinesWrapper
-import org.stepic.droid.features.deadlines.util.getKindOfRecord
+import org.stepik.android.domain.personal_deadlines.model.DeadlinesWrapper
+import org.stepik.android.data.personal_deadlines.getKindOfRecord
 import org.stepic.droid.preferences.SharedPreferenceHelper
 import org.stepic.droid.web.storage.RemoteStorageService
 import org.stepic.droid.web.storage.model.StorageRecord
