+++ app/src/main/java/org/stepik/android/remote/personal_deadlines/mapper/DeadlinesMapper.kt
@@ -1,8 +1,8 @@
 package org.stepik.android.remote.personal_deadlines.mapper
 
 import com.google.gson.GsonBuilder
-import org.stepic.droid.features.deadlines.model.DeadlinesWrapper
-import org.stepic.droid.features.deadlines.util.getKindOfRecord
+import org.stepik.android.domain.personal_deadlines.model.DeadlinesWrapper
+import org.stepik.android.data.personal_deadlines.getKindOfRecord
 import org.stepic.droid.jsonHelpers.adapters.UTCDateAdapter
 import org.stepic.droid.web.storage.model.StorageRecord
 import org.stepic.droid.web.storage.model.StorageRecordWrapped
