+++ app/src/main/java/org/stepik/android/view/personal_deadlines/ui/adapters/EditDeadlinesAdapter.kt
@@ -7,16 +7,16 @@ import android.view.ViewGroup
 import android.widget.TextView
 import kotlinx.android.synthetic.main.view_edit_deadlines_item.view.*
 import org.stepic.droid.R
-import org.stepic.droid.features.deadlines.model.Deadline
+import org.stepik.android.domain.personal_deadlines.model.Deadline
 import org.stepik.android.model.Section
 import org.stepic.droid.ui.util.changeVisibility
 import org.stepic.droid.util.DateTimeHelper
 import java.util.*
 
 class EditDeadlinesAdapter(
-        private val sections: List<Section>,
-        private val deadlines: ArrayList<Deadline>,
-        private val onDeadlineClicked: (Deadline) -> Unit
+    private val sections: List<Section>,
+    private val deadlines: ArrayList<Deadline>,
+    private val onDeadlineClicked: (Deadline) -> Unit
 ): RecyclerView.Adapter<EditDeadlinesAdapter.EditDeadlinesViewHolder>() {
     override fun getItemCount() = sections.size
