+++ app/src/main/java/org/stepik/android/view/personal_deadlines/ui/dialogs/EditDeadlinesDialog.kt
@@ -15,8 +15,8 @@ import com.codetroopers.betterpickers.calendardatepicker.CalendarDatePickerDialo
 import org.stepic.droid.R
 import org.stepic.droid.analytic.Analytic
 import org.stepic.droid.base.App
-import org.stepic.droid.features.deadlines.model.Deadline
-import org.stepic.droid.features.deadlines.model.DeadlinesWrapper
+import org.stepik.android.domain.personal_deadlines.model.Deadline
+import org.stepik.android.domain.personal_deadlines.model.DeadlinesWrapper
 import org.stepik.android.view.personal_deadlines.ui.adapters.EditDeadlinesAdapter
 import org.stepik.android.model.Section
 import org.stepic.droid.util.DateTimeHelper
