+++ app/src/test/java/org/stepic/droid/jsonHelpers/adapters/UTCDateAdapterTest.kt
@@ -3,7 +3,7 @@ package org.stepic.droid.jsonHelpers.adapters
 import com.google.gson.GsonBuilder
 import org.junit.Assert.assertEquals
 import org.junit.Test
-import org.stepic.droid.features.deadlines.model.DeadlinesWrapper
+import org.stepik.android.domain.personal_deadlines.model.DeadlinesWrapper
 import org.stepic.droid.util.toObject
 import java.util.*
